reader_concurrency_semaphore: evict: cancel inactive_read timer
Without canceling the ttl_timer, it may fire after eviction
if the inactive_read is still held by an inactive_read_handle,
causing a double-evict that leads to underflow of
_stats.inactive_reads and _stats.time_based_evictions.

Protect against double evict by checking if the
inactive_read is already unlisted upon entry to evist()
and do nothing in this case.

+++ reader_concurrency_semaphore.cc
@@ -441,7 +441,11 @@ bool reader_concurrency_semaphore::try_evict_one_inactive_read(evict_reason reas
 }
 
 void reader_concurrency_semaphore::evict(inactive_read& ir, evict_reason reason) {
+    if (!ir.is_linked()) {
+        return;
+    }
     auto reader = std::move(ir.reader);
+    ir.ttl_timer.cancel();
     ir.unlink();
     if (auto notify_handler = std::move(ir.notify_handler)) {
         notify_handler(reason);
